fix: resolve feature-stripping build failures and add templ generate to e2e CI
- Add templ generate step to e2e-test job in pipeline.yml (closes #300)
- Extract LinkRelation, HubEntry, RelEntry, Rel(), TitleFromPath(), and
  LinkHeader() into non-demo-gated link_types.go so types survive demo
  stripping (closes #301)
- Move parseFeatureFlag outside demo block and un-gate the setup import
  so the function survives feature stripping (closes #298)
- Gate BreadcrumbsFromLinks call, domain import, and session_settings
  layout logic in handler.go to prevent undefined references after
  stripping

Ref #207
