+++ .github/workflows/pipeline.yml
@@ -144,6 +144,9 @@ jobs:
       - name: Install Playwright browsers
         run: npx playwright install chromium --with-deps
 
+      - name: Generate templ files
+        run: go tool templ generate
+
       - name: Build Go binary
         run: go build -o dothog .
